resolve feedback comments
