+++ examples/ConsoleApplication/appsettings.json
@@ -1,5 +1,7 @@
 {
   "AppName": "Azure App Configuration Console App",
   "Language": "English",
-  "AppConfig:Endpoints": []
+  "AppConfig": {
+    "Endpoints": ["https://my-appconfig.azconfig.io","https://my-another-appconfig.azconfig.io"]
+  }
 }
