--- src/Microsoft.Extensions.Configuration.AzureAppConfiguration/ConfigurationClientFactory.cs
@@ -1,33 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-//
-using Azure.Core;
-using Microsoft.Extensions.Configuration.AzureAppConfiguration.ConfigurationClients;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Microsoft.Extensions.Configuration.AzureAppConfiguration
-{
-    internal class ConfigurationClientFactory : IConfigurationClientFactory
-    {
-        public IConfigurationClient CreateConfigurationClient(string connectionString, AzureAppConfigurationOptions options)
-        {
-            return new FailOverSupportedConfigurationClient(connectionString, options);
-        }
-
-        public IConfigurationClient CreateConfigurationClient(IEnumerable<Uri> endpoints, TokenCredential credential, AzureAppConfigurationOptions options)
-        {
-            if (endpoints == null || endpoints.Count() < 1)
-            {
-                throw new ArgumentNullException(nameof(endpoints));
-            }
-            if (credential == null)
-            {
-                throw new ArgumentNullException(nameof(credential));
-            }
-
-            return new FailOverSupportedConfigurationClient(endpoints, credential, options);
-        }
-    }
-}
